packaging: - Added third optional parameter "nocommit" to script deploy_release-packages.sh. If this parameter is given, the script won't commit the generated package files and keep the temporary directory. Then the script gives directions how to test the building of the packages locally. - deploy_release-packages.sh deletes the temporary directory at start now.
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@2922 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ packaging/deploy_release-packages.sh
@@ -5,18 +5,23 @@ usage() {
 # deploys pacakges
 cat <<EOF
 Usage:
-  $0 <dist-dir> unstable|stable
+  $0 <dist-dir> unstable|stable [nocommit]
+
+If 'nocommit' is specified, the generated files will be kept in the temporary
+directory and not commited to the OpenSuse build service.
+This is useful for building a package locally.
 EOF
-	exit 0
+  exit 0
 }
 
 # parse command line options
 if [ -z "$2" ]; then
-	usage
+  usage
 fi
 
 DIR=$1
 CMD=$2
+NOCOMMIT=$3
 VERSION=`cat $DIR/VER`
 
 TMP_DIR=/tmp/xtreemfs-upload
@@ -24,6 +29,10 @@ TMP_DIR=/tmp/xtreemfs-upload
 if [ ! -f "/usr/bin/osc" ]; then
   echo "osc command not found - please install osc first!"
   exit 1
+fi
+
+if [ -d "$TMP_DIR" ]; then
+  rm -rf "$TMP_DIR"
 fi
 
 cd $DIR
@@ -39,9 +48,12 @@ if [ $CMD == "unstable" ]; then
   # copy all new files, add new and delete old files, check in project
   cp xtreemfs-testing/* $TMP_DIR/home:xtreemfs:unstable/xtreemfs-testing
   osc addremove $TMP_DIR/home:xtreemfs:unstable/xtreemfs-testing/
-  osc ci -m "update" $TMP_DIR/home:xtreemfs:unstable/xtreemfs-testing/
+  if [ -z "$NOCOMMIT" ]; then
+    exit 0
+    osc ci -m "update" $TMP_DIR/home:xtreemfs:unstable/xtreemfs-testing/
   
-  rm -rf $TMP_DIR
+    rm -rf $TMP_DIR
+  fi
   
 elif [ $CMD == "stable" ]; then
     
@@ -64,9 +76,24 @@ elif [ $CMD == "stable" ]; then
   
   cd -
   cd $TMP_DIR/home:xtreemfs
-  osc ci -m "imported xtreemfs $VERSION" xtreemfs-$VERSION
-  cd -
-  
-  rm -rf $TMP_DIR
-  
+  if [ -z "$NOCOMMIT" ]; then
+    osc ci -m "imported xtreemfs $VERSION" xtreemfs-$VERSION
+    cd -
+    
+    rm -rf $TMP_DIR
+  fi
+fi
+
+if [ -n "$NOCOMMIT" ]; then
+  cat <<EOF
+The generated build service files were NOT commited to the OpenSuse build
+service. Instead they are still present in: $TMP_DIR
+
+You can use them to test building a package locally e.g.,
+
+cd $TMP_DIR
+cd home:xtreemfs:unstable/xtreemfs-testing
+# Build the package for openSUSE_11.4, 64 Bit:
+osc build openSUSE_11.4 x86_64 xtreemfs-testing.spec
+EOF
 fi
